+++ src/app/pages/admin/image-list/image-list.component.ts
@@ -1,8 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatTableDataSource } from '@angular/material/table';
 import { ToastrService } from 'ngx-toastr';
 import { Observable } from 'rxjs';
 import { Image } from 'src/app/model/image';
 import { ImageService } from 'src/app/service/image.service';
+import { MatPaginator } from '@angular/material/paginator';
 
 @Component({
   selector: 'app-image-list',
@@ -10,20 +12,45 @@ import { ImageService } from 'src/app/service/image.service';
   styleUrls: ['./image-list.component.scss']
 })
 export class ImageListComponent implements OnInit {
-
+  //Material Paginator
+  @ViewChild ('paginator') paginator: MatPaginator;
+  //
   list$: Observable<Image[]> = new Observable<Image[]>();
   filterKey: string = '';
-
+  //Material Table
+  dataSource: MatTableDataSource<Image>;
+  displayedColumns: string[] = ['name', 'description', 'size', 'type', 'picture','edit']
+  //
   constructor(private iService: ImageService,
-    private toaster: ToastrService) { }
+    private toaster: ToastrService,
+    ) { }
 
   ngOnInit(): void {
     this.list$ = this.iService.list$;
+    //Material Table and paginator
+    this.iService.list$.subscribe(list=>{
+      this.dataSource = new MatTableDataSource(list);
+      this.dataSource.paginator = this.paginator;
+    })
+  }
+  //for Material filter
+  filterTable (filterValue :string) {
+    this.dataSource.filter = filterValue.toLowerCase();
   }
 
   onDelete(image: Image): void {
-    this.iService.remove(image);
+    if(!confirm('Are you sure you want to delete this item?')){
+      return
+    } else {
+      this.iService.remove(image);
     this.toaster.warning('Successfull delete!', 'Deleted', { timeOut: 3000 });
+    }
   }
+
+
+
+
+
+
 
 }
